changed values and moved lib out of public
