+++ components/Layout/index.js
@@ -1,6 +1,6 @@
 import Navigation from "./navigation";
 import { useSession } from "next-auth/react";
-import { StyledMain } from "@/public/lib/styled-components";
+import { StyledMain } from "@/lib/styled-components";
 
 export default function Layout({ children, className }) {
   const { data: sessionData } = useSession();
