+++ components/RoastDetailCard/index.js
@@ -11,7 +11,7 @@ import {
   StyledNumberRating,
   StyledDivButtonWrapper,
   StyledButton,
-} from "@/public/lib/styled-components";
+} from "@/lib/styled-components";
 
 export default function RoastDetailCard({
   onSubmit,
