+++ lib/styled-components.js
@@ -92,9 +92,7 @@ export const StyledDivButtonWrapper = styled.div`
   margin-bottom: 24px;
 `;
 
-// Styled Footer
-
-export const StyledFooter = styled.footer`
+export const StyledFooter = styled.div`
   display: flex;
   justify-content: space-evenly;
   position: fixed;
